backend: listings controller and router
