+++ backend/src/api/games/games.controller.ts
@@ -1,11 +1,8 @@
 import { Assert, type Logger, Result } from "@guillaume-docquier/tools-ts"
-import { z } from "zod"
-import { GameId } from "#api/games/GameId.ts"
+import { type GameId } from "#api/games/GameId.ts"
 import { type PlayerId } from "#api/games/PlayerId.ts"
-import type { ListingsRepository } from "#api/listings/listings.repository.ts"
 import { type LobbyDto, toLobbyDto } from "#api/lobbies/lobbies.controller.ts"
 import { type LobbiesRepository } from "#api/lobbies/lobbies.repository.ts"
-import { GameStatus } from "#api/shared/GameStatus.ts"
 import type { CreateTransaction } from "#lib/db/createDb.ts"
 import type { GamesRepository } from "#lib/db/games/games.repository.ts"
 import type { GameStatesRepository } from "#lib/db/gameStates.repository.ts"
@@ -20,7 +17,6 @@ export class GamesController {
   private readonly createTransaction: CreateTransaction
   private readonly gamesRepository: GamesRepository
   private readonly lobbiesRepository: LobbiesRepository
-  private readonly listingsRepository: ListingsRepository
   private readonly gameStatesRepository: GameStatesRepository
   private readonly gameTicksRepository: GameTicksRepository
   private readonly gamePlayerResourcesRepository: GamePlayerResourcesRepository
@@ -30,7 +26,6 @@ export class GamesController {
     createTransaction,
     gamesRepository,
     lobbiesRepository,
-    listingsRepository,
     gameStatesRepository,
     gameTicksRepository,
     gamePlayerResourcesRepository,
@@ -39,7 +34,6 @@ export class GamesController {
     createTransaction: CreateTransaction
     gamesRepository: GamesRepository
     lobbiesRepository: LobbiesRepository
-    listingsRepository: ListingsRepository
     gameStatesRepository: GameStatesRepository
     gameTicksRepository: GameTicksRepository
     gamePlayerResourcesRepository: GamePlayerResourcesRepository
@@ -48,23 +42,9 @@ export class GamesController {
     this.createTransaction = createTransaction
     this.gamesRepository = gamesRepository
     this.lobbiesRepository = lobbiesRepository
-    this.listingsRepository = listingsRepository
     this.gameStatesRepository = gameStatesRepository
     this.gameTicksRepository = gameTicksRepository
     this.gamePlayerResourcesRepository = gamePlayerResourcesRepository
-  }
-
-  /**
-   * Gets ALL the game listings. This only makes sense until we have real traffic.
-   */
-  public async getListings(): Promise<ListingDto[]> {
-    const getListingsResults = await this.listingsRepository.getListings()
-    if (Result.isFailure(getListingsResults)) {
-      this.logger.error("Could not get game listings, returning empty array", { error: getListingsResults.error })
-      return []
-    }
-
-    return getListingsResults.value
   }
 
   public async startGame({ gameId, playerId }: { gameId: GameId; playerId: PlayerId }): Promise<Result<LobbyDto, string>> {
@@ -126,15 +106,3 @@ export class GamesController {
     return Result.Success(toLobbyDto({ lobbyModel: lobbyResult.value, playerId }))
   }
 }
-
-export type ListingDto = z.infer<typeof ListingDto>
-export const ListingDto = z.object({
-  id: GameId,
-  name: z.string(),
-  nbPlayers: z.number(),
-  nbSeats: z.number(),
-  status: z.enum(GameStatus),
-  createdAt: z.date(),
-  startedAt: z.date().nullable(),
-  endedAt: z.date().nullable(),
-})
